feat : 기능변경 - StudentLecture 레포지토리 파일명 변경... - Lecture에서 생성자 Validate추가 - Professor와 Student의 Count를 매개변수로 받을 때 long타입으로 변경.     - JPA에서 내려주는 타입자체가 long이라서 ㅎ...

+++ src/main/java/com/example/junitexample/domain/entity/Professor.java
@@ -74,7 +74,7 @@ public class Professor extends BaseEntity {
 
     /* ======== 비즈니스로직 ======== */
     // 최초 회원가입시 계정 설정
-    public void defaultSetting(int totalProfessorCount) {
+    public void defaultSetting(long totalProfessorCount) {
         if (totalProfessorCount < 0) {
             throw new IllegalArgumentException("최소값이 0이하는 발생할 수 없습니다.");
         }
