+++ src/main/java/com/example/junitexample/domain/entity/Student.java
@@ -72,7 +72,7 @@ public class Student extends BaseEntity {
 
     /* ======== 비즈니스로직 ======== */
     // 최초 회원가입시 계정 설정
-    public void defaultSetting(int totalStudentCount) {
+    public void defaultSetting(long totalStudentCount) {
         if (totalStudentCount < 0) {
             throw new IllegalArgumentException("최소값이 0이하는 발생할 수 없습니다.");
         }
